Improve the pipeline of some Jenkinsfiles to construct a build version or to release a given version of some projects.
The pipeline for building Docker images are moved into the
releases directory.

Add four new Jenkinsfiles: one for registering a Docker image
for building a Silverpeas project, one for registering a Docker
image for developing on a Silverpeas project, one for preparing
the production-ready Docker image of a version of Silverpeas and
one for creating an IzPack installer of Silverpeas.

The pipeline for releasing a stable version of Silverpeas is
rewritten in the declarative syntax. Instead of doing itself some
jobs (like the Docker images construction or the documentation
generation) it delegates them to external pipelines.