Fix SonarQube Maven execution in pipelines.
In the pipeline dedicated to construct a build version of Silverpeas and
in the pipeline to release a version of Silverpeas,
the execution of the sonarqube analysis fails because the Maven plugin
of SonarQube isn't found. FIx the pb by specifying explicitly the
groupId and artifactId of the SonarQube Maven plugin.